+++ CHANGELOG.md
@@ -13,12 +13,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - `templater.rb` set of tools for creating ERB templates
 - `utils:templatize` method/task to create ERB templates from the command line
+- `common:objects:attach_resources` method
 
 <!-- ### Changed -->
 
 <!-- ### Deprecated -->
 
 <!-- ### Fixed -->
+- fixed typo in `common:objects:move_aos_children_to_parents`
 
 <!-- ### Removed -->